+++ lib/chewy/type/import.rb
@@ -5,13 +5,13 @@ module Chewy
 
       module ClassMethods
         def bulk(options = {})
-          client.bulk options.merge(index: index.index_name, type: type_name)
+          suffix = options.delete(:suffix)
+          client.bulk options.merge(index: index.build_index_name(suffix: suffix), type: type_name)
         end
 
         def import(*args)
           import_options = args.extract_options!
-          bulk_options = import_options.extract!(:refresh).reverse_merge!(refresh: true)
-          identify = {_index: index.index_name, _type: type_name}
+          bulk_options = import_options.extract!(:refresh, :suffix).reverse_merge!(refresh: true)
 
           adapter.import(*args, import_options) do |action_objects|
             payload = {type: self}
@@ -20,9 +20,9 @@ module Chewy
             ActiveSupport::Notifications.instrument 'import_objects.chewy', payload do
               body = action_objects.each.with_object([]) do |(action, objects), result|
                 result.concat(if action == :delete
-                  objects.map { |object| { action => identify.merge(_id: object.respond_to?(:id) ? object.id : object) } }
+                  objects.map { |object| { action => {_id: object.respond_to?(:id) ? object.id : object} } }
                 else
-                  objects.map { |object| { action => identify.merge(_id: object.id, data: object_data(object)) } }
+                  objects.map { |object| { action => {_id: object.id, data: object_data(object)} } }
                 end)
               end
               body.any? ? !!bulk(bulk_options.merge(body: body)) : true