+++ spec/chewy/type/import_spec.rb
@@ -48,11 +48,11 @@ describe Chewy::Type::Import do
     specify do
       dummy_cities.first.destroy
       expect(CitiesIndex.client).to receive(:bulk).with(hash_including(
-        body: [{delete: {_index: 'cities', _type: 'city', _id: dummy_cities.first.id}}]
+        body: [{delete: {_id: dummy_cities.first.id}}]
       ))
       dummy_cities.from(1).each.with_index do |c, i|
         expect(CitiesIndex.client).to receive(:bulk).with(hash_including(
-          body: [{index: {_id: c.id, _index: 'cities', _type: 'city', data: {'name' => "name#{i+1}"}}}]
+          body: [{index: {_id: c.id, data: {'name' => "name#{i+1}"}}}]
         ))
       end
       city.import dummy_cities.map(&:id), batch_size: 1